refactor: separate node-specific connection from server
